Moved some functions to proper classes based on new login flow.
